requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

numpy~=1.25.2
pandas~=2.1.0
setuptools~=68.2.0
selenium~=4.11.2
folium~=0.14.0
flask~=2.3.3
geopy~=2.4.0
psycopg2~=2.9.7
overpy~=0.6
shapely~=2.0.1
requests~=2.31.0
psutil~=5.9.5
geopandas~=0.13.2
seaborn~=0.13.0
matplotlib~=3.8.0
scikit-learn~=1.3.1
xgboost~=2.0.0
nltk~=3.8.1
contractions~=0.1.73
gensim~=4.3.2
googletrans~=4.0.0rc1
openai~=0.28.1
langchain~=0.0.313
mlflow~=2.7.1